requirements.txt
fastapi
uvicorn
requests
pymongo
numpy
python-dotenv
